boil: refactor template logic out of New cmd
