+++ Cargo.toml
@@ -15,9 +15,10 @@ lazy_static = "1.5.0"
 minijinja = { version = "2.12.0", features = ["loader"] }
 nu-ansi-term = "0.50.1"
 reqwest = { version = "0.12.23", features = ["json"] }
-serde = { version = "1.0.219", features = ["derive"] }
-serde_json = "1.0.143"
-toml = "0.9.5"
+serde = { version = "1.0.225", features = ["derive"] }
+serde_json = "1.0.145"
+sqlx = { version = "0.8", features = ["runtime-tokio"] }
+toml = "0.9.6"
 tracing = { version = "0.1.41", features = ["log"] }
 tracing-error = "0.2.1"
 tracing-subscriber = { version = "0.3.20", features = [
