+++ boil/Cargo.toml
@@ -16,6 +16,7 @@ lazy_static = { workspace = true }
 minijinja = { workspace = true }
 nu-ansi-term = { workspace = true }
 serde = { workspace = true }
+sqlx = { workspace = true }
 toml = { workspace = true }
 tracing = { workspace = true }
 tracing-error = { workspace = true }
