+++ boil/src/commands/mod.rs
@@ -1,3 +1,4 @@
+pub(crate) mod add;
 pub(crate) mod list;
 pub(crate) mod new;
 pub(crate) mod test;
